% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfactor.R
\name{check_patterns}
\alias{check_patterns}
\title{Check patterns}
\usage{
check_patterns(f, patterns, ...)
}
\arguments{
\item{f}{A facter}

\item{patterns}{A character vector of regex patterns to be check against the factors levels.}

\item{...}{}
}
\value{

}
\description{
Check patterns
}
