% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfactor.R
\name{xfactor}
\alias{xfactor}
\title{Manipulate factor levels using regex matching.}
\usage{
xfactor(
  x,
  levels = NULL,
  labels = NULL,
  exclude = FALSE,
  reorder = NULL,
  type = "match",
  ...
)
}
\arguments{
\item{x}{A character vector or factor. Usually taking a small number of
distinct values.}

\item{levels}{An optional vector of the unique regular expressions (as
character strings) to be passed to the \code{pattern} argument of
\code{\link{grepl}} or \code{\link{gsub}}. Missing values are allowed.}

\item{labels}{An optional character vector of labels for the levels (in the
same order as levels) For fixed = FALSE this can include backreferences
\code{"\1"} to \code{"\9"} to parenthesized subexpressions of pattern. For perl =
TRUE only, it can also contain \code{"\U"} or \code{"\L"} to convert the rest of the
replacement to upper or lower case and \code{"\E"} to end case conversion. If a
character vector of length 2 or more is supplied, the first element is used
with a warning. If NA, all elements in the result corresponding to matches
will be set to NA.}

\item{exclude}{A vector of regex expressions used to exclude values when
forming the set of levels. Setting \code{exclude = TRUE}, causes all values
not listed in the level or label statements to be excluded.}

\item{reorder}{Sets the final ordering of the factor levels. Can be a
function (or formula), integer, or character variable. Functions (formula)
are passed to the ... argument of `fct_relevel`. Integers are passed to the
`after` argument of `fct_relevel`. Character variables should match one of
the \code{forcats} functions for sorting (`"fct_inorder"`, `fct_infreq`,
`fct_inseq`, `fct_rev`, `fct_shuffle`).}

\item{type}{Controls whether grepl (type = "match") or gsub (type =
"substitute") is used.}

\item{...}{Other arguments passed to \code{\link{grepl}} or
\code{\link{gsub}}, such as ignore.case, perl, fixed.}
}
\value{
\code{xfactor} returns an object of class \code{"factor"} which has a
  set of integer codes the length of x with a \code{"levels"} attribute of
  mode \code{\link{character}} and unique \code{(!\link{anyDuplicated}(.))}
  entries.
}
\description{
Reorder, add to, drop, and/or collapse factor levels.
}
\details{
The encoding of the vector happens as follows. First any values
  matching \code{regex} expressions in \code{exclude} are removed from
  \code{x}. If \code{levels} is a named vector, or \code{labels} is provided,
  existing values for \code{x} are replaced or altered accordingly.

  All \code{labels} are added as factor levels even if no matching values are
  found. If \code{levels} is unnamed and \code{labels} is not provided,
  character strings listed in \code{levels} are moved and/or added to the
  beginning of the final factor levels. If \code{exclude = TRUE} all values
  not listed in \code{levels} or \code{labels} are dropped.

  By default, factor levels are returned in the order of the \code{levels}
  (or \code{labels}) statement. Any values of x that do not match
  \code{levels} are added to the end in their previous order (if \code{x} is
  a factor) or in alphabetical order (if \code{x} is a character).
  
  Level ordering can be controlled further using the \code{reorder} argument.
}
\examples{

x <- c("Man", "Male", "Man", "Lady", "Female")
## Map from 4 different values to only two levels:
xfactor(x, levels = c("Male", "Man" , "Lady", "Female"),
           labels = c("Male", "Male", "Female", "Female"))
xfactor(x, levels = c("^M", "L|F"),
           labels = c("Male", "Female"))
xfactor(x, levels = c(Male = "^M", Female = "L|F"))



}
